+++ src/Ops.DataProvider.SqlServer.Promenade/Migrations/20200316022856_prom_v1.0.0.20.cs
@@ -1,0 +1,23 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
+{
+    public partial class prom_v10020 : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<bool>(
+                name: "IsClosed",
+                table: "Locations",
+                nullable: false,
+                defaultValue: false);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "IsClosed",
+                table: "Locations");
+        }
+    }
+}